mcp-tool-crawler-py/pyproject.toml
[tool.poetry]
name = "mcp-tool-crawler"
version = "0.1.0"
description = "A service for discovering and cataloging Machine Context Protocol (MCP) tools from various sources"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.29.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
RestrictedPython = {version = "^6.2", python = ">=3.9,<3.12"}
aws-lambda-powertools = "^2.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
moto = "^4.2.6"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"

[tool.poetry.group.scripts.dependencies]
requests = "^2.31.0"

[tool.poetry.scripts]
mcp-crawler = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]